requirements.txt
numpy==1.21.2
matplotlib==3.5.0
scipy==1.5.2
pillow==9.1.0
chumpy==0.70
opencv-python-headless==4.5.1.48
opencv-contrib-python-headless==4.5.3.56
open3d==0.12.0
opendr==0.78
pyrender==0.1.45
trimesh==3.8.19
transforms3d==0.3.1
transformers==4.5.1
jax==0.2.12
jaxlib==0.1.65

yacs>=0.1.8
tensorboard>=2.5.0
deprecated
GitPython
tqdm
termcolor
pyyaml
prettytable
networkx
ninja

# vtk 9.1.0
# PyQt5 5.15.6
# mayavi 4.7.4


git+https://github.com/facebookresearch/pytorch3d.git@v0.4.0
git+https://github.com/KailinLi/chamfer_distance.git
git+https://github.com/lixiny/manotorch.git
